+++ src/components/Link/Link.tsx
@@ -3,7 +3,6 @@ import { getCommonProps } from '../../utils';
 import { ComponentProps, ElementType, forwardRef } from 'react';
 import { getLinkVariantClassName, isLinkExternal } from './utils';
 import { LinkVariants } from './types';
-import { Text, TextVariants } from '../Text';
 
 export interface LinkProps extends ComponentProps<'a'> {
   /**
@@ -31,7 +30,7 @@ export interface LinkProps extends ComponentProps<'a'> {
  */
 
 const Link = forwardRef<HTMLAnchorElement, LinkProps>(
-  ({ children, className, element: Element = 'a', variant = LinkVariants.link, href, ...props }, ref) => {
+  ({ children, className, element: Element = 'a', variant = LinkVariants.linkMedium, href, ...props }, ref) => {
     const { className: baseClassName, ...commonProps } = getCommonProps(props, 'Link');
     const classNames = classnames(baseClassName, getLinkVariantClassName(variant), className);
 
@@ -46,7 +45,7 @@ const Link = forwardRef<HTMLAnchorElement, LinkProps>(
         {...(isExternal && Element === 'a' ? { rel: 'noopener noreferrer', target: '_blank' } : {})}
         {...props}
       >
-        <Text variant={variant as unknown as TextVariants}>{children}</Text>
+        {children}
       </Element>
     );
   },