+++ src/patterns/SaleCard/SaleCard.tsx
@@ -86,25 +86,21 @@ const SaleCard = forwardRef<HTMLDivElement, SaleCardProps>(
       <article {...componentProps} className={classes} ref={ref}>
         {imageSrc ? <SeldonImage src={imageSrc} alt={imageAlt} className={`${baseClassName}__image`} /> : null}
         <div className={`${baseClassName}__details`}>
-          <Text variant={TextVariants.badge} className={`${baseClassName}__auction-type`}>
-            {auctionType}
+          <Text variant={TextVariants.labelSmall}>{auctionType}</Text>
+          <Text variant={TextVariants.headingSmall} className={`${baseClassName}__title`}>
+            {titleText}
           </Text>
-          <Text variant={TextVariants.title3}>{titleText}</Text>
           {badgeText && (
-            <Text variant={TextVariants.badge} className={`${baseClassName}__badge`}>
+            <Text variant={TextVariants.labelSmall} className={`${baseClassName}__badge`}>
               {badgeText}
             </Text>
           )}
           <div className={`${baseClassName}__info`}>
-            <Text variant={TextVariants.string2}>{location}</Text>
-            <Text variant={TextVariants.string2}>{date}</Text>
+            <Text variant={TextVariants.labelMedium}>{location}</Text>
+            <Text variant={TextVariants.labelMedium}>{date}</Text>
             {modalButtonText && modalButtonOnClick && (
               <div className={`${baseClassName}__modal-link`}>
-                <Button
-                  onClick={modalButtonOnClick}
-                  variant={ButtonVariants.tertiary}
-                  className={`${baseClassName}__modal-link-button`}
-                >
+                <Button onClick={modalButtonOnClick} variant={ButtonVariants.link}>
                   {modalButtonText}
                 </Button>
               </div>